pyproject.toml
[build-system]
requires = ["setuptools>=65", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "ee-utils"
version = "0.2.0"
description = "Utilities for sampling from Google Earth Engine imagery."
readme = "README.md"
requires-python = ">=3.10"
license = { text = "MIT" }
authors = [{ name = "Jordan Imahori", email = "jordan.imahori@worc.ox.ac.uk" }]
dependencies = [
  "earthengine-api>=0.1.393",
  "numpy>=1.24",
  "pandas>=1.5",
  "requests>=2.31",
  "tqdm>=4.66",
  "pyproj>=3.6",
]
[project.optional-dependencies]
viz = [
  "matplotlib>=3.7",
  "pillow>=10.0",
]
dev = [
  "ruff>=0.4",
  "pytest>=7.4",
]

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]

[tool.ruff]
line-length = 100
target-version = "py310"

[tool.ruff.lint]
select = [
  "E",   # pycodestyle errors
  "F",   # pyflakes
  "I",   # import sorting
  "B",   # flake8-bugbear
  "UP",  # pyupgrade,
  'NPY', # numpy
  'PD',  # pandas
]
ignore = [
    "E501",   # line too long
]